Minimize diff: move as much as possible to `CodeGenRunner`
