+++ language-support/scala/codegen/src/main/scala/com/digitalasset/lf/codegen/lf/DefTemplateWithRecord.scala
@@ -3,7 +3,7 @@
 
 package com.daml.lf.codegen.lf
 
-import com.daml.lf.iface.{DefTemplate, Type, Record}
+import com.daml.lf.iface.{DefTemplate, Record, Type}
 
 final case class DefTemplateWithRecord(
     `type`: Record[Type],
